move private types into separate files
